requirements.txt
asarPy==1.0.1
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-allauth==0.63.3
django-rest-auth==0.9.5
django-rest-knox==4.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
